+++ src/test/java/domain/DealerTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.user.Dealer;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
